backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Background Jobs
celery==5.3.6
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Client
httpx>=0.23.0,<0.26.0

# OpenAI & LLM
openai==1.10.0
anthropic==0.18.0

# LangChain
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph==0.0.20
langfuse==2.10.0

# Replicate (LoRA training/inference)
replicate==0.22.0
Pillow==10.4.0

# Stripe (payments)
stripe==7.11.0

# Storage
boto3==1.34.34  # For R2/S3

# Vector Database
pgvector==0.2.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Rate Limiting
slowapi==0.1.9
